Write tarball contents to a temporary file
Instead of streaming metadata and sequences from a tarball with the
`tarfile.extractfile` method, write the contents of the requested
internal file to a temporary external file. This change allows
downstream processes to read from a file path which is both a more
consistent interface for libraries than the `io.BufferedReader` object
returned by the `tarfile` module and an interface that supports the
multiple passes required by the sanitize metadata script.